+++ hail/hail/src/is/hail/io/bgen/BgenRDDPartitions.scala
@@ -83,7 +83,7 @@ object BgenRDDPartitions extends Logging {
     val getKeysFromFile = StagedBGENReader.queryIndexByPosition(ctx, indexSpec)
 
     nonEmptyFilesAfterFilter.zipWithIndex.map { case (file, fileIndex) =>
-      val nPartitions = math.min(fileNPartitions(fileIndex), file.nVariants).toInt
+      val nPartitions = math.min(fileNPartitions(fileIndex).toLong, file.nVariants).toInt
       val partNVariants: Array[Long] = partition(file.nVariants, nPartitions)
       val partFirstVariantIndex = partNVariants.scan(0L)(_ + _).init
       val partLastVariantIndex = partFirstVariantIndex.lazyZip(partNVariants).map { (idx, n) =>
