+++ hail/hail/test/src/is/hail/types/physical/PContainerTest.scala
@@ -26,7 +26,10 @@ class PContainerTest extends PhysicalTestUtils {
     log.info(s"Testing $data")
 
     val res =
-      Region.containsNonZeroBits(src + sourceType.missingBytesOffset, sourceType.loadLength(src))
+      Region.containsNonZeroBits(
+        src + sourceType.missingBytesOffset,
+        sourceType.loadLength(src).toLong,
+      )
     res
   }
